anv/wsi: Go back to using dummy submits to signal WSI fences
This effectively referts f9a3d9738b12883e268b81731f8e231df3e376c3

In f9a3d9738b12 we attempted to use BO fences to reduce synchronization.
This worked and improved the framerates of some games because we were no
longer synchronizing between the compositor and Vulkan driver constantly.
However, it had the unintended side-effect of waiting on a fence causing
us to wait on whatever work has just been submitted to the GPU which
results in additional GPU idle time.

The previous commit made us start using the RADV trick of not adding the
BO to the GPU exec list when owned by the presentation engine which
should get rid of the extra synchronization between ANV and the
compositor.  This means it should now be safe to go back to the old
mechanism of signaling fences with dummy submits which didn't suffer
from this internal over-synchronization.

Fixes: f9a3d9738b12 "anv: Use BO fences/semaphores for AcquireNextImage"
Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/3039