crocus: Rip out crocus_format_info
This makes sense in iris or ANV where we can assume texture swizzle
everywhere.  In crocus, however, the only piece of supported hardware
with texture swizzle is Haswell (and I guess BSW if we care).  It makes
a lot more sense to just throw around isl_format and use the "real"
format for sampling from things like alpha and luminance textures.  This
patch fixes a pile of GL_ALPHA tests on Haswell which were failing
because we were trying to use ISL_FORMAT_A* with a swizzle.

This does require us to stop advertising a couple luminance formats but
that shouldn't be any great loss.